backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai>=1.0
python-dotenv
tiktoken
numpy
